Generizising character controller example's game.py
