+++ examples/panda3d-character-controller/game.py
@@ -1,7 +1,5 @@
 from panda3d.core import Point3
 from panda3d.core import Vec2, Vec3
-from panda3d.core import NodePath
-from panda3d.core import CollisionSphere
 from panda3d.core import CollisionCapsule
 
 from wecs.core import Component
@@ -9,6 +7,7 @@ from wecs.aspects import Aspect
 from wecs.aspects import factory
 from wecs import panda3d
 from wecs import mechanics
+from wecs.panda3d import aspects
 
 
 # Ignore this for the moment please; It means "This entity's model can be collided into".
@@ -67,7 +66,6 @@ system_types = [
 #   A `non_player_character` is an `avatar` controlled by an `npc_mind`
 #   A `game_map` is a model that you can bump / fall into.
 
-
 game_map = Aspect([panda3d.Position, panda3d.Model, panda3d.Scene, Map],
                   overrides={
                       panda3d.Position: dict(value=factory(lambda:Point3(0, 0, 0))),
@@ -77,97 +75,26 @@ game_map = Aspect([panda3d.Position, panda3d.Model, panda3d.Scene, Map],
 )
 
 
-character = Aspect([panda3d.Clock, panda3d.Position, panda3d.Scene,
-                    panda3d.CharacterController, panda3d.Model],
-                   overrides = {
-                       panda3d.Clock: dict(clock=globalClock),
-                       panda3d.Scene: dict(node=base.render),
-                   }
-)
-def rebecca_bumper():
-    return {
-        'bumper': dict(
-            shape=CollisionSphere,
-            center=Vec3(0.0, 0.0, 1.0),
-            radius=0.7,
-        ),
-    }
-def rebecca_lifter():
-    return {
-        'lifter': dict(
-            shape=CollisionSphere,
-            center=Vec3(0.0, 0.0, 0.25),
-            radius=0.5,
-        ),
-    }
-walking = Aspect([panda3d.WalkingMovement, panda3d.CrouchingMovement, panda3d.SprintingMovement,
-                  panda3d.InertialMovement, panda3d.BumpingMovement, panda3d.FallingMovement,
-                  panda3d.JumpingMovement],
-                 overrides = {
-                     panda3d.InertialMovement: dict(acceleration=30.0, rotated_inertia=0.5),
-                     panda3d.BumpingMovement: dict(solids=factory(rebecca_bumper)),
-                     panda3d.FallingMovement: dict(solids=factory(rebecca_lifter)),
-                     panda3d.JumpingMovement: dict(impulse=factory(lambda: Vec3(0, 0, 6))),
-                 },
-)
-avatar = Aspect([character, walking, mechanics.Stamina],
-                overrides={panda3d.Model: dict(model_name='rebecca.bam')})
-def spectator_bumper():
-    return dict(
-        solids={
-            'bumper': dict(
-                shape=CollisionSphere,
-                center=Vec3(0.0, 0.0, 0.0),
-                radius=0.1,
-            ),
-        },
-    )
-spectator = Aspect([character, panda3d.FloatingMovement, panda3d.BumpingMovement],
-                   overrides={
-                       panda3d.Model: dict(node=factory(lambda:NodePath('spectator'))),
-                       panda3d.BumpingMovement: dict(solids=factory(spectator_bumper)),
-                   },
-)
-
-
-pc_mind = Aspect([panda3d.Input])
-npc_mind = Aspect([panda3d.ConstantCharacterAI])
-
-
-first_person = Aspect([panda3d.FirstPersonCamera],
-                      overrides={panda3d.FirstPersonCamera: dict(camera=base.cam)})
-third_person = Aspect([panda3d.TurntableCamera, panda3d.TurningBackToCameraMovement,
-                       panda3d.CollisionZoom, panda3d.ThirdPersonCamera],
-                      overrides={
-                          panda3d.ThirdPersonCamera: dict(
-                              camera=base.cam,
-                              focus_height=1.8,
-                          ),
-                      },
-)
-
-
-player_character = Aspect([avatar, pc_mind, third_person])
-non_player_character = Aspect([avatar, npc_mind])
-
-
 # Populate the world with the map, the player character, and a few NPCs
 game_map.add(base.ecs_world.create_entity())
-player_character.add(
+
+player_avatar = Aspect([aspects.player_character, mechanics.Stamina])
+
+player_avatar.add(
     base.ecs_world.create_entity(),
     overrides={panda3d.Position: dict(value=Point3(50, 295, 0))},
 )
-non_player_character.add(
+aspects.non_player_character.add(
     base.ecs_world.create_entity(),
     overrides={
         panda3d.Position: dict(value=Point3(50, 300, 0)),
         panda3d.ConstantCharacterAI: dict(
-            move=factory(lambda:Vec3(0.0, 0.25, 0.0)),
+            move=Vec3(0.0, 0.25, 0.0),
             heading=-0.5,
         ),
     },
 )
-non_player_character.add(
+aspects.non_player_character.add(
     base.ecs_world.create_entity(),
     overrides={
         panda3d.Position: dict(value=Point3(60, 300, 0)),
@@ -177,7 +104,7 @@ non_player_character.add(
         ),
     },
 )
-non_player_character.add(
+aspects.non_player_character.add(
     base.ecs_world.create_entity(),
     overrides={panda3d.Position: dict(value=Point3(60, 295, 0))},
 )
